backend/entries/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import DiaryEntryViewSet, TagViewSet, RegisterView

router = DefaultRouter()
router.register(r'entries', DiaryEntryViewSet, basename='entry')
router.register(r'tags', TagViewSet, basename='tag')

urlpatterns = [
    path('auth/register/', RegisterView.as_view(), name='register'),
    path('auth/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('auth/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('', include(router.urls)),
]
